Move common deps to own folder
