+++ apps/common-lib/README.md
@@ -1,0 +1,5 @@
+## common-lib
+
+This folder is for dependencies that are common to all projects – for example, authentication and logging.
+
+Think carefully before adding things to this folder – does every part of what you're adding truly belong here?
