+++ apps/common-lib/src/main/scala/com/gu/typerighter/lib/Loggable.scala
@@ -1,4 +1,4 @@
-package utils
+package com.gu.typerighter.lib
 
 import org.slf4j.{Logger, LoggerFactory}
